currency-converter-a2a-agent/pyproject.toml
[project]
name = "currency-converter-a2a-agent"
version = "0.1.0"
description = "Currency Converter A2A Agent using the official A2A SDK"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.22",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.3.18",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "click>=8.1.8",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
